feat: add Nav component
